fix: refactor and stop form from flickering on sucess
